♻️ Move & rename functions, types and variables
