+++ .gitignore
@@ -51,3 +51,4 @@ release_notes.md
 
 # Testing artifacts
 testarea/
+.vale/
