+++ .vale.ini
@@ -1,0 +1,4 @@
+StylesPath = .vale
+
+[*.md]
+BasedOnStyles = Google
