+++ firefoxversion_test.go
@@ -2,21 +2,14 @@ package main
 
 import (
 	"fmt"
-	"os"
 	"path/filepath"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 )
 
-var homedir string
-
-func init() {
-	homedir, _ = os.UserHomeDir()
-}
-
 func TestFirefoxVersionOfProfile(t *testing.T) {
-	version, err := FirefoxProfileFromPath(filepath.Join(homedir, ".mozilla", "firefox", "667ekipp.default-release")).FirefoxVersion()
+	version, err := NewFirefoxProfileFromPath(filepath.Join(mockedHomedir, ".mozilla", "firefox", "667ekipp.default-release")).FirefoxVersion()
 	assert.NoError(t, err)
 	assert.Equal(t, FirefoxVersion{90, 0}, version)
 }
